requirements.txt
Jinja2==3.0.2
PyYAML==5.3.1
PyPDF2==1.26.0
